Move old prologue/epilogue insertion code to BPatch_addressSpace
